refactor: 💡 change product validations
changed quantity and price validations to support only integer values

+++ app/Concerns/ProductValidations.php
@@ -9,8 +9,8 @@ trait ProductValidations
         return [
             'name' => ['required', 'string', 'min:3', 'max:100'],
             'category' => 'required|exists:categories,name',
-            'quantity' => ['required', 'numeric', 'min:0'],
-            'price' => ['required', 'numeric', 'min:1'],
+            'quantity' => ['required', 'integer', 'min:0'],
+            'price' => ['required', 'integer', 'min:1'],
             'description' => ['required', 'string', 'min:10', 'max:250'],
             'image' => ['required'],
             'image.*' => [
